PP-2446 introducing product client
Similar to other node apps, using the same client approach for interaction with backend microservice. Most of this commit is copy-paste from selfservice's `nuts and bolts` required for REST api calls. (e.g. base_client, pact stuff, loggers, converters etc)

`Product.class` is identical to that from `Selfservice` except in this case it include the `pay_api_key` as this module requires to interact with public_api

Other than that:
- added pact based tests for product client
- included mocha env files for test environment settings