PP-4564: Fix app metrics disabled flag equality
If any environment variable is set for `DIABLE_APP_METRICS` it will
result in `false`, this is because any defined string in JavaScript
converted to Boolean (`!`) will result in `true` - negating this will
always be false.

The current behaviour will disable app metrics if any value is set for
this environment variable (even if it's explicitly false).