PP-5015 - add new custom translation function
i18n’s built in translation function `__()` returns the key
if it can’t find a translation. This is problematic as we show
something gross like 'cardDetails.title' to the user.
So this is a custom function that will check that the string
doesn’t equal the key. And if it does throw an error so the
template fails to render but with a nice helpful error message.