PP-5006 - Making field validations translatable
We use the same validations server and client side so this got a bit tricky, the client-side can rely on the strings object being attached to `window` but the server side can’t. So I decided to add the ability to be able to pass the error message to the check.

The controller can be aware of the language so can pass it using i18n
internally and the client-side checks can get the strings from the
object attached to `window`