PP-5314 - Clear stored reference when passing user to frontend
We were storing in a cookie the value users enter in the reference
field for as long as the cookie persists. This means if you made a
payment link payment and then came back a few days later to do it again
it would show you the reference from the last time, people find the
disconcerting and it’s probably not very useful.

So now when the user is passed from product-ui to frontend the stored
reference is cleared.