PP-5809 Don't log error when error page shown for non-500 errors
We were logging a line at error level whenever an error page was
displayed, but we display the error page for events such as a payment
link not being found.

Only log an error if the status code is 500, not for other statuses.

Also move information in the log line into structured key/values.