PP-6047: Set a few cookies on the products page
Come Chrome 80 (set for release 4th Feb 2020), chrome will treat cookies as
SameSite=Lax by default. (Current default is None.) We need to see if this
change causes cookies set by services to still be present when the user clicks
confirm and the browser redirects the user back to the service's page.

This commit sets a few cookies to test which ones get sent to across sites. This is only
meant to be temporary. Unfortunately this can't be tested locally using `pay
local` as all the hosts are "localhost"!