PP-8943 Add new payment link start page
If a product has type 'ADHOC' or 'AGENT_INITIATED_MOTO' and has the
`new_payment_link_journey_enabled` flag set to true, show the new start
page which displays the payment link description and a "Continue" button
to continue to the next stage of the journey.

If `new_payment_link_journey_enabled` is set to false, continue to
follow the existing payment link journey.

Change how we use Cypress.Cookies.preserveOnce('session') in the
Cypress tests to call it at the beginning of each individual 'it' that
requires the cookies to be carried over. If we call it in 'beforeEach'
then the cookies can be unexpectedly carried over from a different spec
that has run previously.