PP-8982 Add payment link session utilities
We want to index the session for payment links by the product external
id. This is because there is an existing issue that can cause the
session to be interfered with if 2 payment links are opened in different
browser tabs and completed alongside each other.

If a payment link is abandoned, the cookie will also remain in the users
browser for next time they try to complete a payment link if it has not
expired.

Indexing the session by product external ID will ensure at least that a
session cookie remaining from filling in one payment link is not picked
up by a user attempting to complete a completely different payment link
which will more than likely require a different reference and amount.

These utility methods are intended to be used whenever we read/write
from the payment link session.