PP-8966 Accept query parameters for reference and amount
- Accept query parameters 'reference' and 'amount'. If they are present,
set them in the payment link session.
- When the user clicks 'Continue' on the payment link start page, take
them to the next page where a value has not already been provided. If
both reference and amount have been provided or are fixed for the
payment link, 'Continue' takes them to the confirm page.
- If either the reference or amount is not valid, ignore the query
parameter values provided and ask the user to enter them in the UI.