PP-8966 Pass on query params for initial redirect
When a payment link is visited, the GOV.UK router redirects to
`/redirect/:serviceNamePath/:productNamePath`. This controller for this
route resolves the product and redirects to `/pay/:productExternalId`.

After the redirect to `/pay/` is when we decide what page to show the user
depending on the product type and configuration. If the reference and/or
amount are passed as query parameters for a payment link, it makes sense
to put them in the session in the controller for this route. This means
that the `/redirect/` controller needs to preserve the query parameters
when it does the redirect.

As part of this change, the redirect controller now constructs the URL
to redirect itself, rather than using the URL from products. Products
will return either `/pay/:productExternalId` or
`/pay/reference/:productExternalId` as the redirect URL. However, these
routes use exactly the same controller, so instead always redirect to
`/pay/:productExternalId` and we will be able to remove the redundant
`/pay/reference` route in the future.

Replace the integration test that uses supertest with a unit test.
Update Cypress tests that test the start of the payment link journey to
visit the `/redirect/` route.