BAU: categorise CAPTHCA token missing as warning
Card scanning attacks that haven't been built for the payment link pages
generally fail to attempt the CAPTCHA challenge when it's enabled. This
will result in a POST request that doesn't have the token provided by
reCAPTCHA v2.

Currently this scenario will thrown an error which is propagated
through alerting tools. As these are generally high volume attacks this
can introduce a lot of noise.

Treat a missing token as a standard CAPTHCA fail blocking the journey,
warn for analytics but don't raise an error.