PP-8967 Content updates for new payment link journey
Make content updates using the output of the content review.

Change how we decide whether to display the service name (for old payment
link journey and demo payments, etc.) or the payment link name (for the
new payment link journey) in the header. Rather than have a separate
layout, have a conditional in the layout. This means that we will
display the correct header text on the success or failure page for both
old and new payment links.