PP-9476 Fix back links
Fix the back links to make sure that they only ever link to the confirm
page when the user has clicked a change link on the confirm page.
Achieve this by adding a 'change' query parameter to the links from the
confirm page and use the value of this to determine the URL for the
back link.

Use the 'change' query parameter to decide whether the 'Continue' button
on the reference page should redirect to the confirm page, rather than
whether an amount is in the session.

These changes will allow us to add a back link to the confirm page
which will allow the user to step back through the entire journey.