PP-9476 Add back link on confirm page
Add a back link on the confirm page which links to the previous page in
the journey.

Also fix the styling of the back link so that it appears just below the
header banner on all pages, as per the design system guidance.

This previous page is:

The amount page, if the amount is user provided and wasn't provided by
the service in query parameters

ELSE

The reference page, if the reference is user provided and wasn't
provided by the service in query parameters

ELSE

The payment links start page.