PP-9476 Fix redirects when values provided by query params
When the reference or amount are provided by query parameters and the
user clicked the back button to go back to the start page, and then
clicked continue, they were allow to edit the amount or reference.

Fix this so they are either taken to the confirm page, or if one value
was not provided by query params, to the page to provide that.

Fix the behaviour "Continue" button is clicked on the reference page and
only the amount has been provided by query params to redirect to the
confirm page, rather than the amount page.