PP-10028 Fix IEEE 754 floating-point arithmetic errors
Fix IEEE 754 floating-point arithmetic errors that sneaked in when
we switched to the new payment links flow.

Sometimes we need to convert an amount in pounds and pence to an
amount in pence. This might seem simple: multiply it by 100. But
IEEE 754 floating-point arithmetic has other ideas…

100.45 × 100 = 10045              😀
  9.95 × 100 = 994.9999999999999  🙀
131.20 × 100 = 13119.999999999998 😬

Further truncation can mean we end up with an amount that’s a
penny out!

Fix this by avoiding multiplication: instead treat the amount as a
string and remove the decimal point from it.

(Converting an amount in pence to an amount pounds and pence by
dividing by 100 is not as fraught with peril, at least for the
monetary amounts we’re likely to be dealing with: see
https://stackoverflow.com/q/55127530 for details.)