PP-6213 Increase reference max length from 50 characters to 255
Allow the reference number to be 255 characters long to match
what our payments in general allow, rather than just 50.

Also make the field where users enter their reference full-size to
accommodate such jumbo references.

Do not change the validation message just yet: this is because
while this change is being released there will be a period where
some servers are allowing 50 characters and some 255 and it will
confuse a user if they hit a server that only allows 50 and a
server that gives the 255 validation message. So we’ll just pretend
it’s 50 for a little longer.