PP-10136 Fix amount validation for pre-filled payment link
When the amount is specified for the payment link by the service using
query parameters, we do some validation to check it is a valid amount.

However, previously we were incorrectly validating that the amount is
less than or equal to £1000, when this limit should have been £100,000.
We were also ignoring the amount from the query parameters if it was
invalid, and asking users to specify the amount themselves instead.

This has been changed so that we now correctly validate that the amount
is less than or equal to £100,000. We now show an error page instructing
the user to contact the service if either the amount or reference
provided as query parameters are invalid.