PP-10863 Do not autocapitalize ref and omit spellcheck for amount
On the page where the user enters their payment reference, set
autocapitalize="none" because references can have many formats
and capitalising the first letter may not be appropriate.

On the page where the user enters the amount to pay, remove the
spellcheck="false" attribute — nobody should be putting anything
that can be considered a mispelling into this input so it does
not matter whether or not the browser applies spelling checking.