PP-12521 Products UI `request` and `cypress` dev dependency cleanup
We have recently replaced `request` with `axios` in Products UI.

With this change, we are now operating a cleanup of the dev dependencies to
stop `request` showing up in security alerts.

The latest v13 of Cypress no longer needs request-promise-native as a separate
dev dependency (it pulls in @cypress/request).

So we are now:

- removing the override for "request": { "qs": "6.11.1"}

- removing "request-promise-native": "^1.0.9"

- updating cypress to the latest v13.x version as Frontend (13.7.2at time of
  writing)

Further information in JIRA

https://payments-platform.atlassian.net/browse/PP-12521