Support markdown for payment link details
Add a Nunjucks filter to format markdown to HTML

This uses the markdown-it library, which is widely used in MoJ code.

Only support the following:
- paragraphs
- links
- unordered lists
- ordered lists

All other markdown will be ignored, and HTML will be escaped.

Use this Nunjucks filter to format the payment link description.

This implementation re-uses some code from
https://github.com/ministryofjustice/opg-performance-data/ to add
govuk-frontend styling to the rendered HTML elements.