PP-13468 Fix accessibility issues in Payment link confirmation page
With this change, we are fixing two accessibility flaws in the confirmation
page of the Payment Links.

The first is related to the page title, which previously was not very
descriptive and that now explains more clearly what the page is for.

The first issue was the lack of visually hidden element for the second of
two "Change" links.

The element has now been added, so that users of screen readers can be
absolutely confident about the action performed by clicking on that second
link.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13468